feat(core):add option to specify logging path
Add option to specify a logging path. The path specified will have to be a directory, the script will write a file called f5xc-backup-restore.log. If the file exist, we will append to it. Both backup and restore logs will be written to the same file if specified correctly. This option must be specified for the script to work. The log timings are taken based on the timings to make the API calls and file write times are not considered. File sizes are specified in KB.

BREAKING CHANGE:
- Basic sanity checking that directory is specified. No other error handling done. Will need more testing and specific corner cases.