feat(core): Added the capability to perform a force restore
There are situations whereby the tenant is not empty when performing a restore. This is a way to overcome this, although not recommended. In this mode, there is an order of operations that need to be followed during the wipe and restore. Breaking the execution in between will leave the tenant in an unknown state and not advised. There are certain nuances that need to be addressed, in that if there are certain conditions that are undesirable, i.e., an expired SSL certificate, the restore will fail.

BREAKING CHANGE: This will break things, use with caution.

If unsure, we recommend taking a backup and performing a force restore immediately to ensure that the state is consistent. There is no checks to ensure that objects in the backup location is consistent with what is present in the tenant.